Update to Error Prone 2.32.0 (#1037)
Error Prone 2.32.0 requires JDK 17 or higher. We tweak our CI config
accordingly, adding a job testing on the last EP version (2.31.0) that
worked with JDK 11. (We should really tweak our Gradle config to
completely divorce the JDK version used to run Gradle from the JDK
version for the `test` tasks, but that'll have to wait for a PR.)

A few changes to our required job names will be required before this can
land.