Enforce correct type-use annotation locations for nested types (#1045)
Enforce correct type-use annotation locations for nested types as per
JSpecify norms. We enforce type-use annotations to be on the inner class
and raise an error if they are not. For annotations that are both type
use and declaration, we raise an error at an invalid location.

**Current behavior**
```
// All three are ok

@Nullable A.B.C foo1 = null;
A.@Nullable B.C foo2 = null;
A.B.@Nullable C foo3 = null;

```

**New behavior**

```
// BUG: Diagnostic contains: Type-use nullability annotations should be applied on inner class
@Nullable A.B.C foo1 = null;
// BUG: Diagnostic contains: Type-use nullability annotations should be applied on inner class
A.@Nullable B.C foo2 = null;
A.B.@Nullable C foo3 = null;


```

For annotations which are both declaration and annotation and type-use,
only `foo2` throws an error since the location isn't apt for either
scenarios

---------

Co-authored-by: Manu Sridharan <msridhar@gmail.com>