Fix bugs in reading varargs annotations from bytecodes (#1055)
For the case of declaration annotations, we were using `Flags.VARARGS`
to check if a `Symbol` represented a varargs parameter, but this only
works if the method is defined in source code. We add new methods for
varargs parameters that may be defined in bytecodes, and always check
for declaration annotations in such cases.

For type use annotations on parameters, for a method from bytecodes the
annotation info is stored on the method's symbol, not the parameter's.
We borrow some Error Prone code to handle this case.