Fix JarInfer handling of generic types (#1078)
In particular, fix cases where the type of a method parameter either
contains generic type arguments or is itself a type variable. These
fixes are needed for #1072.

We take advantage of the fact that even though `javac` erases generics,
it stores full generic type signatures in class file attributes, which
WALA is able to read. So we can recover generic type information from
these attributes. We needed a bug fix from WALA for this to work, hence
the WALA version bump.

Not every case is supported here (e.g., we haven't tested generic
methods). But these changes are enough to enable #1072 to move forward,
after which we will have a single place to add features and fix bugs.