Add flag to indicate only @NullMarked code should be checked (#1117)
Fixes #574 

We add a new flag `OnlyNullMarked` to indicate that NullAway can only
check code that is `@NullMarked`, and hence it is ok to omit the
(previously required) `AnnotatedPackages` flag. We currently require
_exactly_ one of either `AnnotatedPackages` or `OnlyNullMarked` to be
passed in; if both flags are omitted or both flags are passed, we fail.

As JSpecify and `@NullMarked` become more widely adopted, we may change
the policy to allow neither flag to be passed (as NullAway already
checks `@NullMarked` code out of the box). But for now, we require one
of the flags to be passed, to avoid confusion.