Remove need to use JSpecify's @Nullable annotation (#1142)
Fixes #1139. There are real scenarios where projects may not want to
ship with a JSpecify dependence; see
https://github.com/uber/NullAway/issues/1139#issuecomment-2638705075.
So, we remove any cases where we were specifically checking for or using
JSpecify's `@Nullable` annotation.

Most of the code changes are due to the fact that now, we check if an
annotation is a `@Nullable` annotation using
`Nullness.hasNullableAnnotation`, which requires a `Config` object as a
parameter. So we need to thread a `Config` object as a parameter through
a bunch of methods.