Always acknowledge restrictive annotations in JSpecify mode (#1144)
Previously, we neglected to acknowledge explicit `@Nullable` return
types and `@NonNull` parameter types in `@NullUnmarked` code unless the
`AcknowledgeRestrictiveAnnotations` setting was passed. But, the
JSpecify spec requires acknowledging restrictive annotations in
`@NullUnmarked` code. So now, in JSpecify mode, we always acknowledge
restrictive annotations.

Note that in JSpecify mode, this change will also impact handling of
other "unannotated" code that is not explicitly annotated
`@NullUnmarked`. But, as discussed in #978, we want to switch to
acknowledging restrictive annotations by default anyway. It seems
reasonable to go ahead and make this change in JSpecify mode, to ensure
greater spec compliance.