Better `@MonotonicNonNull` support (#1149)
Fixes #1148 

We add explicit support for any annotation named `@MonotonicNonNull` and
add our own version of the annotation to our annotations package. The
main additional support is that we now reason that once assigned a
non-null value, `@MonotonicNull` fields remain non-null when accessed
from subsequent lambdas, even if the lambdas are invoked asynchronously.