JSpecify: properly handle lambdas in anonymous inner classes (#1165)
Fixes #1156

When checking correctness of lambda overrides we get the type from
`javac`'s inferred type rather than using the enclosing class. Also,
when getting the type of an anonymous class we add some extra assertions
to ensure we find the right `NewClassTree` representing the anonymous
class. I _think_ we'll always get the right tree now, but the asserts
give peace of mind.