Ensure classes null-marked by library model detected in all places (#1197)
Fixes #1194 

Previously we were sloppy about allowing a null `Handler` to be passed
into `CodeAnnotationInfo` and then caching the result. Here we change
the code to always pass a non-null `Handler` in to public APIs of
`CodeAnnotationInfo`, and we avoid caching if the handler is ever `null`
from an internal call.