Prototype implementation of javac plugin to serialize nullness annotations (#1225)
This is a new approach to serializing the annotations from
https://github.com/jspecify/jdk for use with NullAway. The idea is we
build that JDK with this plugin, which will dump out all the discovered
annotation info into JSON files. Then, we write some code to take all
those JSON files and convert them to astubx for NullAway. I went with
JSON instead of directly serializing astubx to iterate more quickly. I
have tested with jspecify/jdk and confirmed I can inject this plugin
into the build and the output JSON files get created. I haven't started
on the code to convert the JSON to astubx yet.

Of course there is much left to do, but I wanted to get this initial
implementation up for review rather than dropping a huge PR at the end
(and also to enable others to help out). We previously thought about
getting out the JDK annotations by parsing the source files (and wrote
some code to do so), but I think this approach will be more robust, as
in the javac plugin we don't need worry about parsing of recent language
features and all symbols are just resolved.

Related to #950 (which probably needs to be updated if we go forward
with this approach)