Updates to deploy releases to Central Portal (#1234)
Fixes #1218. I also updated the Shadow plugin to a more recent version,
as that often has subtle interactions with releasing.

To test, I followed the instructions at
https://github.com/vanniktech/gradle-maven-publish-plugin/releases/tag/0.33.0.
I successfully uploaded a `SNAPSHOT` release with these changes, but we
should cut a normal release after this lands to make sure that works.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Chores**
* Updated Maven Publish Gradle plugin and Shadow plugin to newer
versions.
  * Switched Shadow plugin provider to improve build configuration.
  * Updated publishing properties to enable Maven Central publishing.
  * Adjusted build tasks to use updated Javadoc JAR task names.

<!-- end of auto-generated comment: release notes by coderabbit.ai -->