More generics inference fixes (#1255)
Three fixes:
1. There may be multiple `TypeVariable` objects representing the same
underlying type variable due to differences in annotations (e.g., `T` vs
`@Nullable T`). Switch our maps key'd on `TypeVariable` to use the
underlying `Element` instead, which should uniquely identify the type
variable.
2. Previously when handling constraints, we would treat types like
`@Nullable T` as if they were the same as just `T`; this was wrong. Now,
only treat `TypeVariable` objects with no nullness annotations as actual
type variables whose nullability must be inferred.
3. Before we didn't properly determine whether an `AssignmentTree` was
assigning to a local variable.

For testing, we also add a new config flag
`WarnOnGenericInferenceFailure` to allow for a warning to be emitted
when inference fails for a generic method call. This way, we can test
the expectation that generic inference fails for certain calls. This is
off by default; I'm still unsure how best to expose these
inference-related errors to the user.